backend/requirements.txt
Flask==2.2.5
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
pytest==7.3.1
python-dotenv==1.0.0
flask-cors==4.0.0
Werkzeug==2.2.3
gunicorn==20.1.0
